requirements.txt
asgiref==3.9.1
<<<<<<< HEAD
certifi==2025.7.14
charset-normalizer==3.4.2
coverage==7.3.2
Django==5.2.4
django-cors-headers==4.7.0
django-filter==25.1
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
execnet==2.1.1
factory-boy==3.3.0
Faker==19.6.2
freezegun==1.2.2
idna==3.10
iniconfig==2.1.0
numpy==2.3.1
packaging==25.0
pandas==2.2.2
parameterized==0.9.0
Pillow==10.0.1
pluggy==1.6.0
psutil==5.9.5
PyJWT==2.9.0
pytest==7.4.2
pytest-cov==4.1.0
pytest-django==4.5.2
pytest-mock==3.11.1
pytest-xdist==3.3.1
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
requests==2.32.4
responses==0.23.3
setuptools==80.9.0
six==1.17.0
sqlparse==0.5.3
types-PyYAML==6.0.12.20250516
tzdata==2025.2
urllib3==2.5.0
=======
Django==5.2.4
django-filter==25.1
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
PyJWT==2.9.0
sqlparse==0.5.3
>>>>>>> 7ea824038c121ff7249abb9b6146a19b461a0c27
